chore: move subcmd impls to their definition
